+++ examples/storybook/src/shared/resources/experimentsExample.ts
@@ -3,6 +3,7 @@ export const EXPERIMENTS = [
   { name: 'newFormattingToolbar', scope: 'ricos' },
   { name: 'newSettingsUi', scope: 'ricos' },
   { name: 'tiptapEditor', scope: 'ricos' },
+  { name: 'shouldAddAtomicAsNewBlock', scope: 'ricos' },
   { name: 'fixListLineHeight', scope: 'ricos' },
   { name: 'newVideoVerticalAndSocialModals', scope: 'ricos' },
   { name: 'fixedTabSize', scope: 'ricos' },